Batch and Analysis Added
This release has everything needed (except the data) for paper 2
(optimization paper).

There are some structural changes which allowed the batch running to
work well, but it's largely a rearrangement of code.

The BatchRunner.py script can run multiple files for optimization. The
Analysis.py file is for optimization and analysis.